Restore coming soon page for Albania story release
